+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 📝 ***Note to self**, document the changes so that you can remember what you changed and why.*
 
+### 5/27/2025
+
+#### Added
+- Add support for installing *Nerd fonts*.
+- Add **Cascadia Code** Nerd font.
+
 ### 5/26/2025
 
 #### Added